Add class Component
